Make `jedi:epc` buffer local
As server-pool manages the servers, there is no need to keep
"global" server.  The reason why not making `jedi:epc` buffer
local by default is to stop making a lot of server instances.
This is done by the server-pool now.

Now, The only thing you need to do to setup server per-buffer
is to do something like this in python-mode-hook::

  (set (make-local-variable 'jedi:server-command) ...)
  (set (make-local-variable 'jedi:server-args) ...)

There is no need to call `jedi:start-dedicated-server` in the
hook (it works as before, though).  Note that I can't make
these variable buffer local by default, because it will break
people's setting with `setq`'s.

Other possible way for per-buffer setup is to allow function
for these variables instead of list of strings.  Making `jedi:epc`
buffer local helps in this situation, too.